Rewritten API interface
